Refactor currency representation to use BigDecimal with long storage
- All monetary values in the codebase are now represented using BigDecimal for precision.
- The underlying storage mechanism for currency fields has been switched from double to long type, effectively storing the amounts in cents.
- The front-end display logic has been adjusted to convert the long values from cents to yuan.

+++ boost.common/src/main/java/org/example/common/dataobject/CommodityDO.java
@@ -59,7 +59,7 @@ public class CommodityDO implements Serializable {
     /**
      * Default Unit price of the commodity.
      */
-    private Double unitPrice;
+    private Long unitPrice;
 
     /**
      * Status of the commodity.
