+++ boost.server/src/test/java/org/example/controller/OrderControllerTest.java
@@ -67,9 +67,9 @@ class OrderControllerTest {
 
     @Test
     void testRefundOrder() {
-        when(orderService.refundOrders(any(), any())).thenReturn(new BaseResult<Double>("code", "message", Double.valueOf(0), "requestId"));
+        when(orderService.refundOrders(any(), any())).thenReturn(new BaseResult<Long>("code", "message", Long.valueOf(0), "requestId"));
 
-        BaseResult<Double> result = orderController.refundOrder(new UserInfoModel("sub", "name", "loginName", "aid", "uid", Boolean.TRUE), new RefundOrderParam());
-        Assertions.assertEquals(new BaseResult<Double>("code", "message", Double.valueOf(0), "requestId"), result);
+        BaseResult<Long> result = orderController.refundOrder(new UserInfoModel("sub", "name", "loginName", "aid", "uid", Boolean.TRUE), new RefundOrderParam());
+        Assertions.assertEquals(new BaseResult<Long>("code", "message", Long.valueOf(0), "requestId"), result);
     }
 }
